chore(migrate): remove legacylibrarian and python migration code
The fetch code that was still being used is moved into a separate
source file.

Fixes #5425
